Converted non-C compliant code to standard C. The usage of structures like:
struct x {
  int el1;
  char str[1];
}

and the trick of using a single allocation for str and the structure
itself by allocating sizeof(x) + strlen()-1, are questionable. They
were converted to:

struct x {
  int el1;
  char *str;
}

and there is a single allocation of sizeof(x)+strlen() but then
the str pointer is updated to point to the rest of the data.