Add dependencies from the def files to the libraries that generate them
This avoids build failures in parallel builds, where builds
could fail with this error message:

make[2]: *** No rule to make target `libgnutls-26.def', needed by `all-am'.  Stop.

There is no direct rules that generate it, but it is generated
as a byproduct when building libgnutls.la. By marking the la
file as a dependency, make won't bail out by not finding the file
until that dependency is built, and at that point, the def file
exists.

Signed-off-by: Simon Josefsson <simon@josefsson.org>