Avoid depending on hash order in gdoc.
Previously, gdoc had a hash of regexp replacements for each output
format, and applied the replacements in the order that "keys" returned
for the hash. However, not all orders are safe -- and now that Perl 5.18
randomises hash order per-process, it only worked sometimes!

For example, this order is OK:

'is a #gnutls_session_t structure.'
'\@([A-Za-z0-9_]+)\s*' -> 'is a #gnutls_session_t structure.'
'\%([A-Za-z0-9_]+)' -> 'is a #gnutls_session_t structure.'
'\#([A-Za-z0-9_]+)' -> 'is a @code{gnutls_session_t}  structure.'
'([A-Za-z0-9_]+\(\))' -> 'is a @code{gnutls_session_t}  structure.'

This one, however, winds up producing invalid texinfo:

'is a #gnutls_session_t structure.'
'\%([A-Za-z0-9_]+)' -> 'is a #gnutls_session_t structure.'
'([A-Za-z0-9_]+\(\))' -> 'is a #gnutls_session_t structure.'
'\#([A-Za-z0-9_]+)' -> 'is a @code{gnutls_session_t}  structure.'
'\@([A-Za-z0-9_]+)\s*' -> 'is a  @code{code} {gnutls_session_t}  structure.'

This patch turns the hash into a list, so the replacements will always
be done in the intended order.

Signed-off-by: Adam Sampson <ats@offog.org>