argument descriptions should not end in a dot
When the descrip value for an argument ends in a dot, the rendered
documentation places two dots (for example "specify a password file.."
in srptool(1)).

Most of the descriptions are declared properly (without a trailing
dot), but this patch should clean up the rest.

After this commit, any auto-generated documentation that is committed
to git will probably will also need to be refreshed (or removed from
git entirely and generated from the definitions during build, which
might be cleaner).