srp: Add resistance against guessing usernames
When a client tries to authenticate using an unknown username, instead of
generating a random salt every time, generate the salt based on the
username and a secret seed.

The seed is settable by the application, allowing servers to re-use the
same seed after a restart.

A random seed is generated for each newly allocated SRP server credentials
structure, meaning that applications not using the new API to set the seed
continue to work and gain limited advantage (because they use a different
seed after every restart).

For further information see section 2.5.1.3. in RFC 5054.

Signed-off-by: Attila Molnar <attilamolnar@hush.com>