Overridewq AUTOGEN under --enable-local-libopts only if autogen is not needed.
After commit 6addbc3, specifying --enable-local-libopts unconditionally
replaces the autogen-erated files with their distributed copies, and
substitutes AUTOGEN to false.

The assumption here is that if --enable-local-libopts is not specified,
autogen cannot be installed, and that the distributed copies necessarily
exist.  Neither assumption is always correct.  e.g. someone building a
32-bit copy of GnuTLS from git with a copy of autogen on their system
will have a 64-bit copy of libopts, and a working /usr/bin/autogen,
but not a 32-bit libopts.  Since building autogen depends on Guile,
this is a rather heavyweight pile of gear to require.  (You can force
a successful build in this case, but it requires providing
AUTOGEN=/usr/bin/autogen to make(1), which is distinctly inelegant.)

So fix things so that if any of the distributed copies do not exist, we
do not substitute AUTOGEN, so as to let any copy of autogen that configure
found on the system do its job if necessary, while not forcing the user
to link against the copy of libopts which came with that autogen.

Signed-off-by: Nikos Mavrogiannopoulos <nmav@gnutls.org>