Fix memory leak in wrap_nettle_hash_algorithm().
wrap_nettle_hash_algorithm() leaks an mpz_t if it is called with
pk == GNUTLS_PK_RSA and sig == NULL, in which case it will return
without going through the regular exit path that clears the mpz_t
it allocated at the beginning of the function.  Use the regular
exit path instead to fix this.

This leak can be triggered via calls to
gnutls_pubkey_get_preferred_hash_algorithm().

Signed-off-by: Lennert Buytenhek <buytenh@wantstofly.org>