tests: check gnutls_record_get_state() with DTLS
Since in DTLS we relied on a sliding window to keep track of the
sequence numbers we didn't provide a sensible value to application
via gnutls_record_get_state(). This test makes sure that we report
the "correct" value when asked. Correct being the next number after
the last received packet.