Added auto-detection of getrandom() system call in Linux systems
In addition use getrandom() via the syscall interface if it doesn't
exist in Libc. The reason for the latter is that getrandom() support
for glibc is in limbo for several years, and for auto-detection is that
even if it is going to be present in libc we will not be able to guarrantee
that the system call is available just because it is present in glibc.
For that we detect on initialization whether getrandom() can obtain
random data, and if yes, we continue using that.