gnutls-cli: print 'Handshake was completed'
The change of moving the handshake process as part of the socket
establishment, prevented the text 'Handshake was completed' from
being printed as part of a successful handshake. That message was
used by applications like gnus which use gnutls-cli. This patch
reverts that change and prints that message on successful handshakes.