priorities: Do read crypto policy files with mtime of zero
In a default Fedora Atomic Host installation,
`/etc/crypto-policies/backends/gnutls.config` is a symlink to the
default in `/usr/share/`.  On an OSTree-managed system, files in
`/usr` have an mtime of zero (to help deduplication).

The simple fix here is to still try to read the first time, even if
the file has an mtime of zero.