gnutls_pkcs7_get_embedded_data: added GNUTLS_PKCS7_EDATA_GET_RAW flag
This flag allows the export of the stored embedded data with any wrapping
encoding included. This in particular, it allows to read the data from the
microsoft catalog PKCS#7 structures, which store as embedded data elements
of a SEQUENCE, but only authenticate the inner parts without the bytes
forming the SEQUENCE header.