Restrict the allowed signature algorithms for SKE to the enabled ones
On the other hand, allow the client to sign even with non-enabled
algorithms if necessary. That allows to sign for example with DSA-SHA1
as client even if we do not allow DSA-SHA1 as signature algorithm for
server's certificate. This allows to use a deprecated certificate
without enabling deprecated algorithms globally.

Issue reported by Hubert Kario.