modified the gnutls_certificate_set_key* change
While the change was fully backwards compatible for applications that were
adding a single certificate, and applications that were checking for negative
errors codes, many applications do not. As this may cause incompatibility issues
with software properly utilizing the previously documented API, the change
is reverted, and applications need to explicitly enable a flag (GNUTLS_CERTIFICATE_API_V2)
in the credentials structure for the set_key functions to return an index.