Reverted the behavior of sending a status request extension even without a response
That is, we no longer reply to a client's hello with a status request, with
a status request extension. Although that behavior which was introduced
in 6b76e0c899b1ff08df9bd9b41588f771f050be89 is legal, it creates incompatibility
issues with gnutls 3.3.x branch. That is because versions prior 3.3.26
translates the presence of the extension as a guarrantee that the status
response data will be sent. Even though, that is false assumption we
replicate the previous behavior to allow such clients to connect to
a gnutls 3.5.x server.

Relates !66