tpm: fix handling of keys requiring authorization
There are several problems with the key handling in the tpm code.

The first, and most serious, is that we should make sure we understand
the authorization requirements of a key *before* using it.  The reason
for this is that the TPM has a dictionary attack defence and is
programmed to lock up after a certain number of authorization failures
(which can be very small).  If we try first without authorization, we
may lock up the TPM.  The fix for this is to check whether
authorization is required and supply it before using the key.

Secondly, if the key does require authorization but no password is
supplied we should return immediately, since we know the TPM will give
us an authorization error anyway.

Thirdly, we should unconditionally read the policy of the key rather
than checking if a policy exists: Policies are tied to key objects, so
if there is an old policy in s->tpm_key_policy, but we're creating a
new key, the key it belonged to will be closed, meaning the policy
will be invalid.  Fix this by always setting the policy each time we
get a new key object.

Signed-off-by: James Bottomley <James.Bottomley@HansenPartnership.com>