openpgp.h: all openpgp functionality was marked as deprecated
This is to prevent new applications using that functionality.
As the OpenPGP certificate for HTTPS (or TLS in general) never got
any traction, GnuTLS is the only implementation supporting it,
and the quality of the OpenPGP supporting code is questionable,
we deprecate that code with the intention to drop it completely
when an opportunity is given.

Relates #102

Signed-off-by: Nikos Mavrogiannopoulos <nmav@redhat.com>