gnutls_x509_crt_check_hostname2: do not fallback to CN unconditionally
Do not fallback to checking the CN of a certificate for a hostname
if supported names such as IP addresses were found in gnutls_x509_crt_check_hostname2().
This behavioral change is in order to satisfy the RFC6125 requirement
of not falling back to CN in that case. Reported by Suphannee Sivakorn.

Signed-off-by: Nikos Mavrogiannopoulos <nmav@redhat.com>