OCSP: check the subject public key identifier field to figure issuer
Normally when attempting to match the 'Responder Key ID' in an OCSP response
against the issuer certificate we check (according to RFC6960) against the
hash of the SPKI field. However, in few certificates (see commit:
"added ECDSA OCSP response verification"), that may not be the case. In that
certificate, that value matches the Subject Public Key identifier field
but not the hash.

To account for these certificates, we enhance the matching to also consider
the Subject Public Key identifier field.

Relates: #223

Signed-off-by: Nikos Mavrogiannopoulos <nmav@redhat.com>