gnutls_x509_*_sign: no longer sign with SHA1
Modify the behavior of the functions to sign with an appropriate
to the public key hash algorithm. That although it modifies the
semantics of the functions, it allows them to be useful even after
SHA1 is considered insecure.

In addition to that, the functions which accept a hash algorithm, will
accept a null hash, which instructs the function to select a
reasonable choice.

Signed-off-by: Nikos Mavrogiannopoulos <nmav@redhat.com>