Fix non-null warning
Help the compiler understand the control flow in the MATCH_FUNC and
INVALID_MATCH_FUNC macros.

Because we are using macros, the compiler is not able to correlate the
replaced values of the macro variables to each other yielding non-null
warnings. Introduce a C variable to mimic the macro variable helping
the compiler understanding the control flow.