pkcs11: allow loading trusted modules when pkcs11 was initialized in manual mode
When a PKCS#11 trust module is used in the system, but gnutls_pkcs11_init()
is explicitly called with GNUTLS_PKCS11_FLAG_MANUAL flag, then the PKCS#11
trust store was not loaded, and thus prevent any certificate validation.

This change allows initializing the trust modules only even if generic
PKCS#11 support is disabled by the application.

Relates #316

Signed-off-by: Nikos Mavrogiannopoulos <nmav@redhat.com>