_gnutls_pkcs11_check_init: improved transition between states
The init_level_t for PKCS#11 modules, was incorrectly handled as a
linear state transition, causing few cases in the transition to be
incorrectly handled. Define precisely the state transitions and
enforce them in _gnutls_pkcs11_check_init.

That addresses a regression introduced by the previous state handling
addition, which made impossible to switch from the trusted state to
the all modules.

Signed-off-by: Nikos Mavrogiannopoulos <nmav@redhat.com>