The abi-check target was updated to check against the last tag
As abi-dumper and abi-compliance-checker tools are not reliable when
run across different systems, we now compare the previous tag ABI with the
current compiled library. That is in contrast with the previous behavior
of storing the output files of abi-dumper, which can become obsolete on
a CI update.

That also moves the ABI check only on the CI, and not in the 'make dist' rule
as it takes significant time to run.

This relates to an issue reported against libidn2's use of abi-compliance-checker
but it affects gnutls as they share similar code:
https://gitlab.com/libidn/libidn2/issues/42

Signed-off-by: Nikos Mavrogiannopoulos <nmav@gnutls.org>